Use PkpForm instead of CounterReportForm in CounterReportsEditModal
